feature:producer

+++ com/ordersystem/model/AbstractOrder.java
@@ -3,11 +3,11 @@ import java.util.Map;
 import java.util.HashMap;
 
 public class AbstractOrder {
-    protected int id;
+    protected long id;
     protected OrderStatus status;
     protected Map<MenuItem, Integer> items;
 
-    public AbstractOrder(int id) {
+    public AbstractOrder(long id) {
         this.id = id;
         this.items = new HashMap<>();
         this.status = OrderStatus.PENDING;
